ml-pipeline/requirements.txt
pandas==1.3.3
numpy==1.19.5
statsmodels==0.12.2
streamlit==1.2.0
plotly==5.1.0
tensorflow-cpu==2.7.0
scikit-learn==0.24.2
